\name{simulated_data}
\alias{simulated_data}
\docType{data}
\title{Dataset simulated for toy example}
\description{
A toy dataset simulated for demonstration for the \code{splmm} function. 
}
\usage{data(simulated_data)}
\format{
  \describe{
   \item{y}{Response variable.}
   \item{x}{Fixed-effects design matrix.}
   \item{z}{Random-effects design matrix}
   \item{grp}{Subject ID. }
  }
}


\examples{
data(simulated_data)
}
\keyword{datasets}
